embassy: Refactor module structure to remove kitchen-sink `util`.
