+++ embassy-hal-common/src/lib.rs
@@ -3,6 +3,7 @@
 // This mod MUST go first, so that the others see its macros.
 pub(crate) mod fmt;
 
+pub mod drop;
 pub mod interrupt;
 mod macros;
 pub mod peripheral;
