+++ embassy-nrf/src/time_driver.rs
@@ -2,9 +2,9 @@ use core::cell::Cell;
 use core::sync::atomic::{compiler_fence, AtomicU32, AtomicU8, Ordering};
 use core::{mem, ptr};
 use critical_section::CriticalSection;
+use embassy::blocking_mutex::CriticalSectionMutex as Mutex;
 use embassy::interrupt::{Interrupt, InterruptExt};
 use embassy::time::driver::{AlarmHandle, Driver};
-use embassy::util::CriticalSectionMutex as Mutex;
 
 use crate::interrupt;
 use crate::pac;
