+++ embassy/src/channel/mod.rs
@@ -1,0 +1,4 @@
+//! Async channels
+
+pub mod mpsc;
+pub mod signal;
