+++ embassy/src/waitqueue/mod.rs
@@ -1,0 +1,5 @@
+//! Async low-level wait queues
+
+#[cfg_attr(feature = "executor-agnostic", path = "waker_agnostic.rs")]
+mod waker;
+pub use waker::*;
